requirements.txt
argparse
nose
pyzmq>=16,<18
ujson
unittest2
mock
flake8
